requirements.txt
streamlit
librosa
numpy
openai-whisper
imageio-ffmpeg
python-docx
spleeter
tensorflow